Simplifies Cottontail DB database driver based on recent refactoring activities.
