+++ gradle.properties
@@ -7,6 +7,7 @@ version_javacv=1.5.9
 version_javalin=6.1.3
 version_javalinopenapi=6.1.3
 version_javalinssl=6.1.3
+version_jdbc_postgres=42.7.3
 version_jline=3.23.0
 version_junit=5.10.1
 version_junit_platform=1.10.1
