+++ vitrivr-engine-module-cottontaildb/src/main/kotlin/org/vitrivr/engine/plugin/cottontaildb/descriptors/scalar/ScalarDescriptorReader.kt
@@ -10,6 +10,7 @@ import org.vitrivr.engine.core.model.descriptor.vector.VectorDescriptor.Companio
 import org.vitrivr.engine.core.model.metamodel.Schema
 import org.vitrivr.engine.core.model.query.Query
 import org.vitrivr.engine.core.model.query.bool.SimpleBooleanQuery
+import org.vitrivr.engine.core.model.query.fulltext.SimpleFulltextQuery
 import org.vitrivr.engine.core.model.retrievable.Retrieved
 import org.vitrivr.engine.core.model.types.toValue
 import org.vitrivr.engine.plugin.cottontaildb.*
@@ -31,29 +32,10 @@ class ScalarDescriptorReader(field: Schema.Field<*, ScalarDescriptor<*>>, connec
      *
      * @param query The [Query] to execute.
      */
-    override fun query(query: Query): Sequence<ScalarDescriptor<*>> {
-        require(query is SimpleBooleanQuery<*>) { "Query of type ${query::class} is not supported by ScalarDescriptorReader." }
-
-        /* Prepare query. */
-        val cottontailQuery = org.vitrivr.cottontail.client.language.dql.Query(this.entityName)
-            .select(RETRIEVABLE_ID_COLUMN_NAME)
-            .select(DESCRIPTOR_ID_COLUMN_NAME)
-            .select(VALUE_ATTRIBUTE_NAME)
-            .where(Compare(Column(this.entityName.column(VALUE_ATTRIBUTE_NAME)), query.operator(), Literal(query.value.toCottontailValue())))
-
-        /* Execute query. */
-        return this.connection.client.query(cottontailQuery).asSequence().map {
-            this.tupleToDescriptor(it)
-        }
-    }
-
-    /**
-     * Executes the provided [Query] and returns a [Sequence] of [Retrieved]s that match it.
-     *
-     * @param query The [Query] to execute.
-     */
-    override fun queryAndJoin(query: Query): Sequence<Retrieved> {
-        TODO("Not yet implemented")
+    override fun query(query: Query): Sequence<ScalarDescriptor<*>> = when (query) {
+        is SimpleFulltextQuery -> this.queryFulltext(query)
+        is SimpleBooleanQuery<*> -> this.queryBoolean(query)
+        else -> throw UnsupportedOperationException("The provided query type ${query::class.simpleName} is not supported by this reader.")
     }
 
     /**
@@ -72,6 +54,47 @@ class ScalarDescriptorReader(field: Schema.Field<*, ScalarDescriptor<*>>, connec
             is IntDescriptor -> IntDescriptor(retrievableId, descriptorId, tuple.asInt(VALUE_ATTRIBUTE_NAME)?.toValue() ?: throw IllegalArgumentException("The provided tuple is missing the required field '$VECTOR_ATTRIBUTE_NAME'."))
             is LongDescriptor -> LongDescriptor(retrievableId, descriptorId, tuple.asLong(VALUE_ATTRIBUTE_NAME)?.toValue() ?: throw IllegalArgumentException("The provided tuple is missing the required field '$VECTOR_ATTRIBUTE_NAME'."))
             is StringDescriptor -> StringDescriptor(retrievableId, descriptorId, tuple.asString(VALUE_ATTRIBUTE_NAME)?.toValue() ?: throw IllegalArgumentException("The provided tuple is missing the required field '$VECTOR_ATTRIBUTE_NAME'."))
+        }
+    }
+
+    /**
+     * Executes a [SimpleFulltextQuery] and returns a [Sequence] of [ScalarDescriptor]s.
+     *
+     * @param query The [SimpleFulltextQuery] to execute.
+     * @return [Sequence] of [ScalarDescriptor]s.
+     */
+    private fun queryFulltext(query: SimpleFulltextQuery): Sequence<ScalarDescriptor<*>> {
+        val cottontailQuery = org.vitrivr.cottontail.client.language.dql.Query(this.entityName)
+            .select("*")
+            .fulltext(VALUE_ATTRIBUTE_NAME, query.value.value, "score")
+
+        if (query.limit < Long.MAX_VALUE) {
+            cottontailQuery.limit(query.limit)
+        }
+
+        /* Execute query. */
+        return this.connection.client.query(cottontailQuery).asSequence().map {
+            this.tupleToDescriptor(it)
+        }
+    }
+
+    /**
+     * Executes a [SimpleBooleanQuery] and returns a [Sequence] of [ScalarDescriptor]s.
+     *
+     * @param query The [SimpleBooleanQuery] to execute.
+     * @return [Sequence] of [ScalarDescriptor]s.
+     */
+    private fun queryBoolean(query: SimpleBooleanQuery<*>): Sequence<ScalarDescriptor<*>> {
+        /* Prepare query. */
+        val cottontailQuery = org.vitrivr.cottontail.client.language.dql.Query(this.entityName)
+            .select(RETRIEVABLE_ID_COLUMN_NAME)
+            .select(DESCRIPTOR_ID_COLUMN_NAME)
+            .select(VALUE_ATTRIBUTE_NAME)
+            .where(Compare(Column(this.entityName.column(VALUE_ATTRIBUTE_NAME)), query.operator(), Literal(query.value.toCottontailValue())))
+
+        /* Execute query. */
+        return this.connection.client.query(cottontailQuery).asSequence().map {
+            this.tupleToDescriptor(it)
         }
     }
 }
